refactor: StatusType 및 StatusTypeEnum의 경로를 공통 경로로 수정 #193
