+++ frontend/src/pages/createdMentoring/components/MentoringApplicationStatus/MentoringApplicationStatus.tsx
@@ -1,6 +1,9 @@
 import styled from '@emotion/styled';
 
-import { StatusTypeEnum, type StatusType } from '../../types/statusType';
+import {
+  StatusTypeEnum,
+  type StatusType,
+} from '../../../../common/types/statusType';
 
 import type { myTheme } from '../../../../common/types/theme';
