+++ frontend/src/pages/createdMentoring/components/PhoneNumber/PhoneNumber.tsx
@@ -1,6 +1,9 @@
 import styled from '@emotion/styled';
 
-import { StatusTypeEnum, type StatusType } from '../../types/statusType';
+import {
+  StatusTypeEnum,
+  type StatusType,
+} from '../../../../common/types/statusType';
 
 interface PhoneNumberProps {
   status: StatusType;
